refactor : Register 패키징 구조 수정
