+++ backend/src/main/java/com/dajava/backend/domain/register/dto/admin/AdminLoginRequest.java
@@ -1,0 +1,4 @@
+package com.dajava.backend.domain.register.dto.admin;
+
+public class AdminLoginRequest {
+}
